+++ src/clients/IGiftCardDataSource.ts
@@ -1,6 +1,6 @@
 import { IGiftCard, IGiftCardPayload } from '@/models/IGiftCard';
 
 export interface IGiftCardDataSource {
-  fetchProducts(page: number): Promise<void>;
+  fetchProducts(page: number): Promise<{ data: unknown; status: number }>;
   emitGiftCard(payload: IGiftCardPayload): Promise<IGiftCard>;
 }