+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # CHANGELOG
 
+## versão [1.1.3](https://github.com/l3co/kudos-jaya/compare/1.1.1...1.1.3)
+
+### fixed
+
+- Token validation while building home app
+
+## versão [1.1.2](https://github.com/l3co/kudos-jaya/compare/1.1.1...1.1.2)
+
+### fixed
+
+- Fix app home build when there is an error
+
+## versão [1.1.1](https://github.com/l3co/kudos-jaya/compare/1.1.0...1.1.1)
+
+### fixed
+
+- Minimum amount to give kudos
+
 ## versão [1.1.0](https://github.com/l3co/kudos-jaya/compare/1.0.0...1.1.0)
 
 ### modified